backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
eth-account
eth-utils
web3
python-dotenv
